Upped version for release again since I forgot to commit the last release upgrade to source control. Needed to update some broken things from reverting version to 8

+++ app/src/main/java/org/morningstarcc/app/fragments/EventFragment.java
@@ -1,13 +1,13 @@
 package org.morningstarcc.app.fragments;
 
+import android.os.Handler;
+import android.os.Looper;
 import android.support.v4.widget.SwipeRefreshLayout;
 
 import com.android.volley.Response;
 import com.j256.ormlite.android.apptools.OpenHelperManager;
 import com.j256.ormlite.dao.Dao;
 
-import org.joda.time.DateTime;
-import org.joda.time.Days;
 import org.morningstarcc.app.R;
 import org.morningstarcc.app.activities.EventActivity;
 import org.morningstarcc.app.adapters.DatabaseItemAdapter;
@@ -19,6 +19,7 @@ import org.morningstarcc.app.libs.DateUtils;
 
 import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.Calendar;
 import java.util.List;
 import java.util.Locale;
 
@@ -46,64 +47,50 @@ public class EventFragment extends RecyclerFragment<Event> {
     }
 
     private void refreshData(){
+        try {
+            Database database = OpenHelperManager.getHelper(getActivity(), Database.class);
+            Dao<Event, Integer> dao2 = database.getDao(Event.class);
+            final DatabaseItemAdapter adapter = getAdapter(fetchData(dao2));
 
-        getActivity().runOnUiThread(new Runnable() {
-            @Override
-            public void run() {
-
-                Database database = OpenHelperManager.getHelper(getActivity(), Database.class);
-                try {
-                    Dao<Event, Integer> dao2 = database.getDao(Event.class);
+            new Handler(Looper.getMainLooper()).post(new Runnable() {
+                @Override
+                public void run() {
                     recyclerView.removeAllViews();
-                    recyclerView.setAdapter(getAdapter(fetchData(dao2)));
-                } catch (SQLException e) {
-                    e.printStackTrace();
+                    recyclerView.setAdapter(adapter);
                 }
-
-
-            }
-        });
-
-
+            });
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
     }
 
     @Override
     public void customInit(){
-
         enableSwipe(new SwipeRefreshLayout.OnRefreshListener() {
             @Override
             public void onRefresh() {
                 DataService.updateEvents(getActivity(), new Response.Listener<Void>() {
                     @Override
                     public void onResponse(Void response) {
-
-                        //Delete data
                         Database database = OpenHelperManager.getHelper(getActivity(), Database.class);
-                        Dao<Event, Integer> dao;
+
                         try {
-                            dao = database.getDao(Event.class);
-                            DateTime currentDate = new DateTime();
+                            Dao<Event, Integer> dao = database.getDao(Event.class);
                             for (Event item : dao) {
-                                DateTime eventDate = new DateTime(DateUtils.getFullDate(item.eventstarttime));
-                                int days = Days.daysBetween(currentDate.toDateMidnight(), eventDate.toDateMidnight()).getDays();
-                                if (days < 0) {
-                                    dao.delete(item);
-                                }
+                                Calendar date = Calendar.getInstance();
+                                date.setTime(DateUtils.getFullDate(item.eventstarttime));
+
+                                if (Calendar.getInstance().after(date)) dao.delete(item);
                             }
-                        } catch (SQLException e) {
-                            e.printStackTrace();
                         } catch (Exception e) {
                             e.printStackTrace();
                         }
 
                         refreshData();
                         endRefresh();
-
                     }
                 });
-
             }
         });
-
     }
 }
